·

+++ upgrade.sh
@@ -5,15 +5,21 @@
 
 
 function Diy_Part1() {
-  echo "正在执行：给源码增加定时更新固件插件和设置插件和ttyd成默认自选"
-  rm -rf "$HOME_PATH/package/luci-app-autoupdate"
-  git clone https://github.com/shidahuilang/luci-app-autoupdate $HOME_PATH/package/luci-app-autoupdate > /dev/null 2>&1
-  [[ -f "$BUILD_PATH/AutoUpdate.sh" ]] && cp -Rf $BUILD_PATH/AutoUpdate.sh $BASE_PATH/bin/AutoUpdate.sh
-  [[ -f "$BUILD_PATH/replace.sh" ]] && cp -Rf $BUILD_PATH/replace.sh $BASE_PATH/bin/replace.sh
-  sed  -i  's/ luci-app-ttyd//g' $HOME_PATH/target/linux/*/Makefile
-  sed  -i  's/ luci-app-autoupdate//g' $HOME_PATH/target/linux/*/Makefile
-  sed -i 's?DEFAULT_PACKAGES +=?DEFAULT_PACKAGES += luci-app-autoupdate luci-app-ttyd?g' $HOME_PATH/target/linux/*/Makefile
-  [[ -d $HOME_PATH/package/luci-app-autoupdate ]] && echo "增加定时更新插件成功"
+  if [[ -f "$BUILD_PATH/AutoUpdate.sh" ]]; then
+    echo "正在执行：给源码增加定时更新固件插件和设置插件和ttyd成默认自选"
+    rm -rf "$HOME_PATH/package/luci-app-autoupdate"
+    git clone https://github.com/shidahuilang/luci-app-autoupdate $HOME_PATH/package/luci-app-autoupdate > /dev/null 2>&1
+    [[ ! -d "$BASE_PATH/usr/bin" ]] && mkdir $BASE_PATH/usr/bin
+    cp -Rf $BUILD_PATH/AutoUpdate.sh $BASE_PATH/usr/bin/AutoUpdate
+    cp -Rf $BUILD_PATH/replace.sh $BASE_PATH/usr/bin/replace
+    chmod 777 $BASE_PATH/usr/bin/AutoUpdate $BASE_PATH/usr/bin/replace
+    sed  -i  's/ luci-app-ttyd//g' $HOME_PATH/target/linux/*/Makefile
+    sed  -i  's/ luci-app-autoupdate//g' $HOME_PATH/target/linux/*/Makefile
+    sed -i 's?DEFAULT_PACKAGES +=?DEFAULT_PACKAGES += luci-app-autoupdate luci-app-ttyd?g' $HOME_PATH/target/linux/*/Makefile
+    [[ -d $HOME_PATH/package/luci-app-autoupdate ]] && echo "增加定时更新插件成功"
+  else
+    echo "没发现AutoUpdate.sh文件存在，不能增加在线升级固件程序"
+  fi
 }
 
 function GET_TARGET_INFO() {
@@ -70,8 +76,11 @@ function GET_TARGET_INFO() {
 	;;
 	esac
 	
-	AutoUp_Ver="$BASE_PATH/bin/AutoUpdate.sh"
-	[[ -f ${AutoUp_Ver} ]] && export AutoUpdate_Version=$(egrep -o "V[0-9].+" $BASE_PATH/bin/AutoUpdate.sh | awk 'END{print}')
+	if [[ -f "$BASE_PATH/usr/bin/AutoUpdate" ]]; then
+	  export AutoUpdate_Version=$(egrep -o "Version=V[0-9]\.[0-9]" $BASE_PATH/usr/bin/AutoUpdate |cut -d "=" -f2 | sed 's/^.//g')
+	else
+	  export AutoUpdate_Version="7.1"
+	fi
 	export In_Firmware_Info="$BASE_PATH/bin/openwrt_info"
 	export Github_Release="${Github}/releases/tag/AutoUpdate"
 	export Openwrt_Version="${SOURCE}-${TARGET_PROFILE}-${Upgrade_Date}"
@@ -83,28 +92,28 @@ function GET_TARGET_INFO() {
 }
 
 function Diy_Part2() {
-	GET_TARGET_INFO
-	cat >${In_Firmware_Info} <<-EOF
-	Github=${Github}
-	Author=${Author}
-	Library=${Library}
-	Warehouse=${Warehouse}
-	SOURCE=${SOURCE}
-	LUCI_EDITION=${LUCI_EDITION}
-	DEFAULT_Device=${TARGET_PROFILE}
-	Firmware_SFX=${Firmware_SFX}
-	TARGET_BOARD=${TARGET_BOARD}
-	CURRENT_Version=${Openwrt_Version}
-	LOCAL_CHAZHAO=${LOCAL_CHAZHAO}
-	CLOUD_CHAZHAO=${CLOUD_CHAZHAO}
-	Download_Path=/tmp/Downloads
-	Version=${AutoUpdate_Version}
-	API_PATH=/tmp/Downloads/Github_Tags
-	Github_API1=${Github_API1}
-	Github_API2=${Github_API2}
-	Github_Release=${Github_Release}
-	Release_download=${Release_download}
-	EOF
+GET_TARGET_INFO
+cat >${In_Firmware_Info} <<-EOF
+Github=${Github}
+Author=${Author}
+Library=${Library}
+Warehouse=${Warehouse}
+SOURCE=${SOURCE}
+LUCI_EDITION=${LUCI_EDITION}
+DEFAULT_Device=${TARGET_PROFILE}
+Firmware_SFX=${Firmware_SFX}
+TARGET_BOARD=${TARGET_BOARD}
+CURRENT_Version=${Openwrt_Version}
+LOCAL_CHAZHAO=${LOCAL_CHAZHAO}
+CLOUD_CHAZHAO=${CLOUD_CHAZHAO}
+Download_Path=/tmp/Downloads
+Version=${AutoUpdate_Version}
+API_PATH=/tmp/Downloads/Github_Tags
+Github_API1=${Github_API1}
+Github_API2=${Github_API2}
+Github_Release=${Github_Release}
+Release_download=${Release_download}
+EOF
 }
 
 function Diy_Part3() {
